+++ lib/persistence.dart
@@ -1,4 +1,8 @@
 library persistence;
 
-export 'src/hive_persistence.dart';
-export 'src/persistence.dart';
+export 'src/persistence_interface.dart';
+export 'src/persistence_subject_interface.dart';
+export 'src/implementations/hive_persistence.dart';
+export 'src/implementations/volatile_persistence.dart';
+export 'src/subjects/persistence_subject.dart';
+export 'src/subjects/volatile_subject.dart';