+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.0 - 2021.04.26
+
+ - Included behavior subject and yours dependencies to local directory;
+ - Added IPersistenceSubject and your basic implementations;
+ - Added IPersistence and your basic implementations;
+
 ## [1.0.0+1] - 24.01.21
 
 * First official version from this package
+